pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wireguard-gui"
dynamic = ["version"]
authors = [
  { name="Rick Dawson", email="dawsora@gmail.com" },
]
description = "A package for setting up and managing WireGuard VPN networks through a web-based GUI"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "alembic>=1.13.1",
  "arrow>=1.3.0",
  "attrs>=23.2.0",
  "binaryornot>=0.4.4",
  "blinker>=1.7.0",
  "certifi>=2024.2.2",
  "cffi>=2.0.0",
  "chardet>=5.2.0",
  "charset-normalizer>=3.3.2",
  "cli-ui>=0.19.0",
  "click>=8.1.7",
  "colorama>=0.4.6",
  "colorclass>=2.2.2",
  "contextlib2>=21.6.0",
  "cryptography>=46.0.3",
  "docopt>=0.6.2",
  "flask>=3.0.3",
  "flask-cors>=6.0.2",
  "flask-login>=0.6.3",
  "flask-migrate>=4.0.7",
  "flask-sqlalchemy>=3.1.1",
  "greenlet>=3.0.3",
  "idna>=3.7",
  "itsdangerous>=2.2.0",
  "jinja2>=3.1.3",
  "location>=0.0.7",
  "mako>=1.3.3",
  "markdown-it-py>=4.0.0",
  "markupsafe>=2.1.5",
  "mdurl>=0.1.2",
  "packaging>=24.0",
  "pastel>=0.2.1",
  "pillow>=10.3.0",
  "platformdirs>=4.2.1",
  "psutil>=5.9.8",
  "pycparser>=2.22",
  "pygments>=2.17.2",
  "pyopenssl>=24.1.0",
  "pypng>=0.20220715.0",
  "pyroute2>=0.7.12",
  "python-dateutil>=2.9.0.post0",
  "python-slugify>=8.0.4",
  "pyyaml>=6.0.1",
  "qrcode>=7.4.2",
  "requests>=2.31.0",
  "rich>=13.7.1",
  "schema>=0.7.5",
  "segno>=1.6.1",
  "setuptools>=69.5.1",
  "six>=1.16.0",
  "sqlalchemy>=2.0.29",
  "tabulate>=0.9.0",
  "terminaltables>=3.1.10",
  "text-unidecode>=1.3",
  "typing-extensions>=4.11.0",
  "unidecode>=1.3.8",
  "urllib3>=2.2.1",
  "werkzeug>=3.0.2",
  "wheel>=0.43.0",
  "win-inet-pton>=1.1.0"
]

[project.urls]
Homepage = "https://github.com/radawson/wireguard-gui"
Issues = "https://github.com/radawson/wireguard-gui/issues"

[tool.hatch.version]
path = "src/gui/__init__.py"
